requirements.txt
chromadb==0.3.29
deep_translator==1.11.4
deepl==1.16.1
Flask==2.3.2
Flask_SocketIO==5.3.4
gevent==22.10.2
Markdown
openai==1.9.0
Pillow
protobuf
PyYAML
Requests
tiktoken
transformers
yeelight
sentence_transformers
gevent-websocket
pymdown-extensions
pygments
pydantic[email]
multipart
aiofiles
fastapi
colorlog